requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
configparser==6.0.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
fonttools==4.46.0
frozendict==2.3.8
getFamaFrenchFactors==0.0.5
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pandas-datareader==0.10.0
parso==0.8.3
patsy==0.5.3
peewee==3.17.0
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.40
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
requests-futures==1.0.1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.0
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
wcwidth==0.2.10
webencodings==0.5.1
yahooquery==2.3.7
yfinance==0.2.31
zipp==3.17.0
